Expose the packaging as a plugin hook (#952)
Resolves #951. This will allow having PEP-517/518 for the end users.
